refactor(runbotics-orchestrator-ui): :recycle: Refactored AddEmailSubscriptionComponent
